refactor runtime into modular structure
